% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestralDistrib.R
\name{AncestralDistrib}
\alias{AncestralDistrib}
\title{Ancestors distribution}
\usage{
AncestralDistrib(anc.niche, predictors, bioclim.past.data,
  bioclim.ext = "tif", start.boost = 0, boost.step = 0.1)
}
\arguments{
\item{anc.niche}{object of class "AncestralNiche"}

\item{predictors}{predictors to use. A vector obtained with
\code{\link{SuggestVariables}}.}

\item{bioclim.past.data}{a character string with the path to geospatial
layers in other timeframe (BIOCLIM or similar variables). The format of
layers must be compatible with \code{stack} and \code{extract} methods
of \code{raster} package.}

\item{bioclim.ext}{a character string with the file extention of geospatial
layers given in bioclim.past.data argument. By default is "tif" (geotif files).}

\item{start.boost}{a number to begin doing factor boosting if there are no
compatible conditions in bioclim.past.data.}

\item{boost.step}{a number to sequential increase of boosting level}
}
\value{
an object of class AncestralDistrib
}
\description{
Ancestors distribution
}
