% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestralDistribSave.R
\name{AncestralDistribSave}
\alias{AncestralDistribSave}
\title{Save ancestor distribution to CSV and GeoTiff files}
\usage{
AncestralDistribSave(anc.distrib, output.path)
}
\arguments{
\item{anc.distrib}{an object of "AncestralDistrib" class.}

\item{output.path}{a character string with path to save produced files.}
}
\value{
null
}
\description{
Save ancestor distribution to CSV and GeoTiff files
}
