% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncestralNiche.R
\name{AncestralNiche}
\alias{AncestralNiche}
\title{Predict ancestral niche}
\usage{
AncestralNiche(trees, best.tree, predictors.values, outgroup = F, nodes,
  trees.sample = 100, trees.burnin = 20, do.lambda = T, do.models = F)
}
\arguments{
\item{trees}{object of class "multiPhylo".}

\item{best.tree}{object of class "phylo".}

\item{predictors.values}{a named numeric vector of predictor values obtained
by \code{\link{SelectPredVals}}.}

\item{outgroup}{vector of phylogenetic tips for outgroup. Not necessary.}

\item{nodes}{internal nodes to estimate states (combined characters
or numbers).}

\item{trees.sample}{trees sample fraction (how many trees to analyze).}

\item{trees.burnin}{trees burn-in percentage (how many trees to drop before
sampling a fraction).}

\item{do.lambda}{a flag whether to estimate phylogenetic sygnal of
predictors on trees.}

\item{do.models}{a flag whether to estimate best fitting evolutionary model
for obtaining ancestral values for predictors.}
}
\value{
object of class "AncestalNiche"
}
\description{

}
