% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recline.R
\name{GetFraction}
\alias{GetFraction}
\title{Get fraction of phylogenetic trees}
\usage{
GetFraction(trees, burnin = 20, sample = 100)
}
\arguments{
\item{trees}{a "multiPhylo" object with multiple trees.}

\item{burnin}{the percentage of starting trees to discard.}

\item{sample}{a number of trees to sample randomly after burn-in.}
}
\value{
a "multiPhylo" object with new trees selection.
}
\description{
Perform burn-in and get random sapmle inside set of trees.
}
\examples{
# get random 100 trees after discarding first 20\%
forest.sample <- GetFraction(forest)
}
