% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recline.R
\name{GetPredictorsValues}
\alias{GetPredictorsValues}
\title{Get predictor values for each occurence point.}
\usage{
GetPredictorsValues(occurence.points, bioclim.data, bioclim.ext = "tif")
}
\arguments{
\item{occurence.points}{a character string with the path to CSV file
containing occurence points. Necessary columns in the CSV file are:
Type, ID, Long, Lat, Clade.}

\item{bioclim.data}{a character string with the path to geospatial layers
(BIOCLIM or similar variables). The format of layers must be compatible with
\code{stack} and \code{extract} methods of \code{raster} package.}

\item{bioclim.ext}{a character string with the file extention of geospatial
layers given in bioclim.data argument. By default is "tif" (geotif files).}
}
\value{
list with two collections,
\code{$raw} (extracted BIOCLIM data for each occurence point) and
\code{$values} (including columns ID, Type, and Clade of occurence.points +
data from \code{$raw} collection).
}
\description{
Predictor values is the values of each geospatial layer
(BIOCLIM) for each occurence point of the species of interest.
}
