% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recline.R
\name{MultiplePhylosig}
\alias{MultiplePhylosig}
\title{Multiple phylogenetic signal test}
\usage{
MultiplePhylosig(trees, testdata)
}
\arguments{
\item{trees}{a "multiPhylo" object with multiple trees.}

\item{testdata}{a vector with continuous trait with names matching
trees tips}
}
\value{
data.frame.
}
\description{
Multiple phylogenetic signal test
}
