\name{rcn-package}
\alias{rcn-package}
\alias{rcn}
\docType{package}
\title{
\packageTitle{rcn}
}
\description{
\packageDescription{rcn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rcn}
\packageIndices{rcn}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rcn}

Maintainer: \packageMaintainer{rcn}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
