\name{recline-package}
\alias{recline-package}
\alias{recline}
\docType{package}
\title{
\packageTitle{recline}
}
\description{
\packageDescription{recline}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{recline}
\packageIndices{recline}
The package provides several functions for easy and convenient obtaining climatic niche of any species or group of specimens by a collection of their points of occurrence. Climatic niche space is constructed by PCA of predictors values of each point derived from BIOCLIM variables. It also contains functions for making a reconstruction of the ancestral climatic niche by given ultrametric phylogenetic tree and for estimation of the probable distribution area of the ancestor by predictors values restored from PCA.
}
\author{
\packageAuthor{recline}

Maintainer: \packageMaintainer{recline}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
