% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recline.R
\name{SelectPredVals}
\alias{SelectPredVals}
\title{Select predictor values}
\usage{
SelectPredVals(vars, predvals)
}
\arguments{
\item{vars}{a character vector with names of selected variables.}

\item{predvals}{predictor values obtained with
\code{\link{GetPredictorsValues}}.}
}
\value{
A named numeric vector.
}
\description{
Extract predictors values only for selected variables
after selection done by \code{\link{SuggestVariables}}.
}
