% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recline.R
\name{SuggestVariables}
\alias{SuggestVariables}
\title{Suggest predictors variables}
\usage{
SuggestVariables(predictors.values)
}
\arguments{
\item{predictors.values}{numeric matrix or data.frame of predictor values
for used occurence points.}
}
\value{
a character vector with the names of suggested variables.
}
\description{
Suggest predictors variables for prediction of ancestral
niche & distibution by PCA and Broken Stick model. Suggested variables
contain the maximum information about studied species.
}
\examples{
predictors <- GetPredictorsValues(occurence.points, bioclim.present)
predictors.num <- predictors$values[1:3112,4:ncol(predictors$values)]
predictors.use <- SuggestVariables(predictors.num)
predictors.selected <- SelectPredVals(predictors.use, predictors$values)
}
